amod_encoder/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "amod-encoder"
version = "0.1.0"
description = "ROI-agnostic reproduction of the AMOD encoding-model pipeline (Jang et al.)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "David", email = "placeholder@example.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "nibabel>=5.0",
    "nilearn>=0.10",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "typer[all]>=0.9",
    "rich>=13.0",
    "h5py>=3.9",
    "statsmodels>=0.14",
    "pydantic>=2.0",
    "matplotlib>=3.8",
    "Pillow>=10.0",
]

[project.optional-dependencies]
gpu = ["torch>=2.0", "timm>=1.0"]
extractors = ["torch>=2.0", "timm>=1.0", "opencv-python>=4.8", "Pillow>=10.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
]
all = ["amod-encoder[gpu,extractors,dev]"]

[project.scripts]
amod-encoder = "amod_encoder.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
